requirements.txt
audiocraft
Fabric
matplotlib
muselsl
numpy
Pillow
Pillow
pyaudio
pylsl
Requests
scikit_learn
scipy
torch
torchaudio
tqdm
websockets
accelerate
diffusers
torch
transformers
hydra-core
matplotlib
pandas
tqdm
Pillow
ftfy
regex
clip @ git+https://github.com/openai/CLIP.git